启用 -Wall -Wextra 编译器警告并修复所有告警
在根 CMakeLists.txt 中启用 -Wall -Wextra -Wno-unused-parameter，
设置 C11 标准。修复自有代码中所有 -Wsign-compare 警告。

+++ libcskburn_serial/src/core.c
@@ -138,7 +138,7 @@ try_sync(cskburn_serial_device_t *dev, int timeout)
 		} else if (ret != -ETIMEDOUT) {
 			return ret;
 		}
-	} while (TIME_SINCE_MS(start) < timeout);
+	} while (TIME_SINCE_MS(start) < (uint32_t)timeout);
 	return -ETIMEDOUT;
 }
