+++ libcskburn_usb/src/core.c
@@ -28,7 +28,7 @@ cskburn_usb_exit(void)
 static libusb_device *
 find_device(int16_t bus, int16_t address)
 {
-	int64_t count = 0;
+	ssize_t count = 0;
 	libusb_device **devs;
 
 	if ((count = libusb_get_device_list(NULL, &devs)) < 0) {
@@ -38,7 +38,7 @@ find_device(int16_t bus, int16_t address)
 	libusb_device *found = NULL;
 	struct libusb_device_descriptor desc = {0};
 
-	for (size_t i = 0; i < count; i++) {
+	for (ssize_t i = 0; i < count; i++) {
 		if (libusb_get_device_descriptor(devs[i], &desc) < 0) {
 			continue;
 		}
